+++ editors/garden-mode.el
@@ -329,8 +329,8 @@ the user entering a value in the *garden* buffer."
        'symbols)
      . font-lock-keyword-face)
     (,(regexp-opt
-       '("dbg" "int_to_string" "path_exists" "print" "println"
-         "shell" "working_directory")
+       '("dbg" "path_exists" "print" "println"
+         "shell" "string_repr" "working_directory")
        'symbols)
      . font-lock-builtin-face)